# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create a Razorpay order to unlock Days 2-28
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Payments not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a Razorpay payment signature and unlock full access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Payment verified; user unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /users/{id}/profile:
    get:
      operationId: getPublicProfile
      tags: [community]
      summary: Get a user's public profile (only if they opted in)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: Not found or profile is private
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Me:
      type: object
      properties:
        id:
          type: integer
        email:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        currentDay:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        streakCount:
          type: integer
        streakFreezes:
          type: integer
        weeklyXp:
          type: integer
        isPublicProfile:
          type: boolean
        hideFromLeaderboard:
          type: boolean
        anonymousMode:
          type: boolean
        emailRemindersEnabled:
          type: boolean
        lastActiveDate:
          type: ["string", "null"]
        theme:
          type: string
        hasPaid:
          type: boolean
        isGrandfathered:
          type: boolean
        hasAccess:
          type: boolean
          description: True if the user can access all 28 days (paid or grandfathered).
        accessibleThroughDay:
          type: integer
          description: Highest day number the user can currently open (1 for free preview, 28 if entitled).
        onboardingCompleted:
          type: boolean
        workContext:
          type: ["string", "null"]
        programTitle:
          type: ["string", "null"]
        programFocus:
          type: ["string", "null"]
        dailyMinutes:
          type: integer
          description: Target minutes the learner wants to spend per day (10, 20, or 40). Drives lesson depth and day-completion requirements.
        field:
          type: ["string", "null"]
          description: The domain the learner wants to apply AI to (e.g. marketing, engineering).
        experienceLevel:
          type: ["string", "null"]
          description: Self-reported AI experience (beginner, some, advanced).
        goal:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - currentDay
        - dailyMinutes
        - xp
        - level
        - streakCount
        - streakFreezes
        - weeklyXp
        - isPublicProfile
        - hideFromLeaderboard
        - anonymousMode
        - emailRemindersEnabled
        - theme
        - hasPaid
        - isGrandfathered
        - hasAccess
        - accessibleThroughDay
        - onboardingCompleted
        - createdAt
...
    PaymentOrder:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: integer
          description: Amount in the smallest currency unit (paise).
        currency:
          type: string
        razorpayKeyId:
          type: string
      required:
        - orderId
        - amount
        - currency
        - razorpayKeyId
    PaymentVerifyInput:
      type: object
      properties:
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string
      required:
        - razorpayOrderId
        - razorpayPaymentId
        - razorpaySignature
...
    Achievement:
      type: object
      properties:
        code:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        unlockedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - code
        - title
        - description
        - icon
...
    PublicProfile:
      type: object
      properties:
        userId:
          type: integer
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        level:
          type: integer
        xp:
          type: integer
        streakCount:
          type: integer
        completedDays:
          type: integer
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
        joinedAt:
          type: string
          format: date-time
      required:
        - userId
        - level
        - xp
        - streakCount
        - completedDays
        - badges
        - joinedAt
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...